+++ modules/nixosModules/hosts/tower/configuration.nix
@@ -1,6 +1,10 @@
-{
+{config, ...}: {
   nix.settings.secret-key-files = ["/var/lib/nix-signing/tower-1.sec"];
-  networking.hostName = "tower";
+  networking = {
+    hostName = "tower";
+    firewall.allowedUDPPorts = [config.services.tailscale.port];
+  };
+  services.tailscale.enable = true;
   users.users.kage = {
     isNormalUser = true;
     description = "Kage";